+++ packages/react/src/patterns/Navigation/Sidebar/Searchbar/index.tsx
@@ -1,9 +1,10 @@
 import { ButtonHTMLAttributes } from "react"
 
+import { Shortcut } from "@/ui/Shortcut"
+
 import { F0Icon } from "../../../../components/F0Icon"
 import { Search } from "../../../../icons/app"
 import { cn, focusRing } from "../../../../lib/utils"
-import { Shortcut } from "@/ui/Shortcut"
 
 interface SearchBarProps extends ButtonHTMLAttributes<HTMLButtonElement> {
   placeholder: string
@@ -21,7 +22,7 @@ export function SearchBar({
       <button
         onClick={onClick}
         className={cn(
-          "mb-[calc(0.75rem-1px)] flex w-full cursor-pointer items-center justify-between rounded bg-f1-background-inverse-secondary p-1.5 text-f1-foreground-secondary ring-1 ring-inset ring-f1-border-secondary transition-all hover:ring-f1-border-hover",
+          "mb-[calc(0.75rem-1px)] flex w-full cursor-pointer items-center justify-between rounded bg-f1-background-inverse-secondary dark:bg-f1-background-tertiary p-1.5 text-f1-foreground-secondary ring-1 ring-inset ring-f1-border-secondary transition-all hover:ring-f1-border-hover",
           focusRing()
         )}
         type="button"